pyproject.toml
[project]
name = "inspect-assist"
version = "0.1.0"
description = "Context-aware LLM assistant for industrial thermal inspection systems"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "openai>=1.50,<2",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",
    "jinja2>=3.1,<4",
    "python-multipart>=0.0.12",
    "httpx>=0.27,<1",
    "Pillow>=10.4,<11",
    "pyyaml>=6.0,<7",
    "structlog>=24.4,<25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.7,<1",
    "mypy>=1.12,<2",
    "httpx>=0.27,<1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
